[ENH] merge of `sktime` fork into `pykalman` - `numpy 2` compatibility, documentation, test refactor (#120)
This PR merges the `sktime` fork back into `pykalman`. All changes were
authored by myself.

Contents:

* documentation improvements
* `numpy 2` compatibility patch, fixes
https://github.com/pykalman/pykalman/issues/118
* tests refactored to be `pytest` idiomatic
* linting to be in line with `sktime` pre-commits

The contents are equivalent with the current state of the `sktime` fork,
with the exception of:

* removing conditional test skips and `sktime` dependencies

The changes introduce a dependency on `scikit-base`, used for its
dependency checking utilities.

`scikit-base` has no dependencies itself, so should be fine - it will
also be used for registry capabilities later.