[DOC] Migrating Examples to Notebooks: plot_em.ipynb EM estimation of model parameters with a Kalman Filter (#164)
# Examples to Notebook Migration

As mentioned in https://github.com/pykalman/pykalman/issues/129 and
following a brief discussion with @fkiraly on Discord, I am initiating
the process of migrating examples in the documentation from Python files
to a notebook.

If I understand this one file, I can work on migrating all the examples
into notebook examples and adding them into the docs correctly.