[BUG] fix time-varying transition covariance being overwritten by initial matrix in standard Kalman filter (#166)
In the filtering of standard Kalman Filter, the transition covariance
and the observation covariance performs an in-place operation, resulting
in always using the covariances for the first timestep.

Namely, in `standard.py`:
```py
transition_covariance = _last_dims(transition_covariance, t - 1)
# <a few lines later>
observation_covariance = _last_dims(observation_covariance, t)
```
This is a bug that has been reported ever since **2015**, yet still was
not fixed in later versions.
https://github.com/pykalman/pykalman/issues/44

This is a really important bug because all results obtained using this
package with time-varying transition-covariance are flawed. Thankfully
the fix is really simple and I strongly suggest to integrate this pull
request asap. Many thanks for this awesome repository!