[DOC] README - credit to duckworthd (#176)
The contributors pane of GitHub now acknowledges me as the first
contributor to `pykalman`, likely due to the large number of (many
smaller or maintenance focused) PR. Imo this unfairly diminishes the
contribution of @duckworthd. Unfortunately, this is done by a black-box
algorithm in GitHub, so we have no direct influence on the ranking
afaik.

To offset this, I think we should add a sentence "originally created by
@duckworthd" to the readme intro - this PR adds such a sentence,
prominently placed.